Switch `CreateWorktreeResponse::worktree_id` to be a `i32`
Apparently all IDs in `sqlx` on the server are represented as i32s.

+++ zed-rpc/src/rest.rs
@@ -2,6 +2,6 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Serialize, Deserialize)]
 pub struct CreateWorktreeResponse {
-    pub worktree_id: u64,
+    pub worktree_id: i32,
     pub rpc_address: String,
 }
